Cleanup test_emscripten_async_wget2
- Name the next more meaningfully
- Merge into a single file
- Remove tabs
- Fix fomatting
- Put trivial methods in class
